+++ apps/playground/src/index.css
@@ -4,6 +4,11 @@ html, body {
     overflow: hidden;
 }
 
+iframe {
+    width: 100%;
+    height: 800px;
+}
+
 .banner {
     position: absolute;
     top: 50%;
